login and register page, with spare
added the login page, register page and the forgot password if the user logs in and doesn't remember their password which will get their email and send the password (this is just in case can be fully coded later). Main parts which is the login and register with the css separating the styling and the code.